+++ src/App.tsx
@@ -4,7 +4,6 @@ import ViewPortContainer from 'components/@commons/ViewPortContainer';
 import ErrorFallback from 'error/ErrorFallback';
 import { Provider } from 'jotai';
 import HomeIndex from 'pages/HomeIndex';
-import KakaoAuthPage from 'pages/KakaoAuthPage';
 import SelectWeekPage from 'pages/SelectWeekPage';
 import SignupPage from 'pages/SignupPage';
 import AddGroupPage from 'pages/admin/AddGroupPage';
@@ -12,6 +11,7 @@ import ApplicationClosePage from 'pages/admin/ApplicationClosePage';
 import ApplicationOpenPage from 'pages/admin/ApplicationOpenPage';
 import ApplyPage from 'pages/alba/ApplyPage';
 import InvitedPage from 'pages/alba/InvitedPage';
+import KakaoAuthPage from 'pages/auth/KakaoAuthPage';
 import LogoutOnlyPrivate from 'privateRoutes/LogoutOnlyPrivate';
 import UserTypePrivate from 'privateRoutes/UserTypePrivate';
 import { ErrorBoundary } from 'react-error-boundary';
