+++ src/pages/SelectWeekPage/AdminDetailSection/index.tsx
@@ -2,11 +2,11 @@ import SubmitButton from 'components/@commons/SubmitButton';
 import Text from 'components/@commons/Text';
 import { DailyWorkersTable } from 'components/DailyWorkersTable';
 import Loader from 'components/Suspenses/Loader';
-import { useGetDailyWorkers } from 'hooks/SchedulePage/fetch';
-import { useGetApplyStatus } from 'hooks/SelectWeekPage/fetch';
-import useWeeklyDetail from 'hooks/SelectWeekPage/useWeeklyDetail';
 import useWeekSelector from 'hooks/useWeekSelector';
 import { useAtomValue } from 'jotai';
+import { useGetDailyWorkers } from 'pages/SchedulePage/hooks/fetch';
+import { useGetApplyStatus } from 'pages/SelectWeekPage/hooks/fetch';
+import useWeeklyDetail from 'pages/SelectWeekPage/hooks/useWeeklyDetail';
 import { Suspense } from 'react';
 import { stringDateMove } from 'utils/dateToString';
 import { selectedWeekAtom } from '../states';
