+++ src/pages/admin/ApplicationClosePage/hooks/fetch.ts
