+++ src/pages/admin/ApplicationOpenPage/SetTimeTemplateSection/index.tsx
@@ -5,8 +5,8 @@ import SubmitButton from 'components/@commons/SubmitButton';
 import Text from 'components/@commons/Text';
 import { AddButton } from 'components/@commons/icons/buttons';
 import SetTimeTemplateSkeleton from 'components/Suspenses/PageSkeletons/SetTimeTemplateSkeleton';
-import { useGetOpenTemplate } from 'hooks/admin/ApplicationOpenPage/fetch';
-import useTimeTemplate from 'hooks/admin/ApplicationOpenPage/useTimeTemplate';
+import { useGetOpenTemplate } from 'pages/admin/ApplicationOpenPage/hooks/fetch';
+import useTimeTemplate from 'pages/admin/ApplicationOpenPage/hooks/useTimeTemplate';
 import { Suspense } from 'react';
 import { myTheme } from 'styles/myTheme';
 import { stringDateMoveKor } from 'utils/dateToString';
